# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AaabPashi Web API
  description: |
    AaabPashi is a comprehensive water management and agricultural monitoring platform API.

    The API provides authentication via phone number and OTP, field management, geospatial 
    data processing, water release scheduling, and multilingual support.

    ## Authentication

    This API supports two authentication methods:

    ### 1. JWT-based Authentication (Primary)
    JWT-based authentication with phone number and OTP verification. 
    Authentication tokens are stored in HTTP-only cookies for security.

    ### 2. API Key Authentication (Alternative)
    API key authentication for programmatic access, third-party integrations, 
    and administrative operations. API keys can be provided in headers:
    - `Authorization: Bearer <api-key>`
    - `X-API-Key: <api-key>`

    ### Authentication Priority
    1. API key authentication is checked first
    2. If API key fails, JWT authentication is attempted
    3. If both fail, request is rejected with 401 Unauthorized

    ## Rate Limiting

    OTP endpoints are rate-limited to one request per phone number every 5 minutes.

    ## External Services

    The API integrates with:
    - Veevo SMS API for OTP delivery
    - Google Translate API for multilingual support
    - Open-Meteo API for weather data
    - Google Earth Engine API for satellite imagery
    - Custom Python APIs for geospatial processing

    ## Multi-Platform Sync

...
  version: 1.0.0
...
servers:
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /{locale}/api/auth/login:
    post:
      tags:
        - Authentication
      summary: Send login OTP
      description: Send OTP to existing user for login
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
            enum: [en, ur]
          description: Language locale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid phone number format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Phone number not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: OTP already sent, please wait
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /{locale}/api/auth/verifyotp-login:
    post:
      tags:
        - Authentication
      summary: Verify login OTP
      description: Verify OTP and complete user login
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
            enum: [en, ur]
          description: Language locale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPLoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Authentication cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid OTP or phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/api-keys:
    get:
      tags:
        - API Key Management
      summary: List all API keys
      description: Retrieve all API keys (admin only)
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: List of API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKeyInfo"
                  count:
                    type: integer
                    example: 2
        "401":
          description: Unauthorized - Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - API Key Management
      summary: Create new API key
      description: Create a new API key (admin only)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/api-keys/{id}:
    get:
      tags:
        - API Key Management
      summary: Get API key details
      description: Get specific API key details (admin only)
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID
      responses:
        "200":
          description: API key details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ApiKeyInfo"
        "401":
          description: Unauthorized - Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - API Key Management
      summary: Update API key
      description: Update API key properties (admin only)
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApiKeyRequest"
      responses:
        "200":
          description: API key updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: API key updated successfully
                  data:
                    $ref: "#/components/schemas/ApiKeyInfo"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - API Key Management
      summary: Delete API key
      description: Delete an API key (admin only)
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID
      responses:
        "200":
          description: API key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: API key deleted successfully
        "401":
          description: Unauthorized - Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only cookie
  schemas:
...
    LoginRequest:
      type: object
      required:
        - phoneNumber
      properties:
        phoneNumber:
          type: string
          example: "+923001234567"
          description: Registered phone number
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "OTP sent successfully"
        phoneNumber:
          type: string
          example: "+923001234567"
        id:
          type: string
          example: "64a1b2c3d4e5f6789012345"
        division:
          type: string
          example: "Kasur"
        name:
          type: string
          example: "John Doe"
    VerifyOTPLoginRequest:
      type: object
      required:
        - phoneNumber
        - otp
      properties:
        phoneNumber:
          type: string
          example: "+923001234567"
        otp:
          type: string
          example: "1234"
...
    CreateApiKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "My App"
          description: Human-readable name for the API key
        userId:
          type: string
          example: "user123"
          description: Associated user ID (optional)
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, admin, delete]
          example: ["read", "write"]
          description: Array of permissions for the API key
        expiresInDays:
          type: integer
          minimum: 1
          example: 30
          description: Expiration in days (optional)
    UpdateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          example: "Updated App Name"
          description: Human-readable name for the API key
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, admin, delete]
          example: ["read", "write", "admin"]
          description: Array of permissions for the API key
        isActive:
          type: boolean
          example: true
          description: Whether the API key is active
        expiresInDays:
          type: integer
          minimum: 1
          example: 60
          description: Expiration in days (0 to remove expiration)
    ApiKeyInfo:
      type: object
      properties:
        id:
          type: string
          example: "abc123def456"
          description: Unique API key identifier
        name:
          type: string
          example: "My App"
          description: Human-readable name
        userId:
          type: string
          example: "user123"
          description: Associated user ID (if any)
        permissions:
          type: array
          items:
            type: string
          example: ["read", "write"]
          description: Array of permissions
        isActive:
          type: boolean
          example: true
          description: Whether the key is active
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
          description: Creation timestamp
        lastUsed:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
          description: Last usage timestamp
        expiresAt:
          type: string
          format: date-time
          example: "2024-02-01T00:00:00.000Z"
          description: Expiration timestamp (if any)
    CreateApiKeyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "API key created successfully"
        data:
          type: object
          properties:
            id:
              type: string
              example: "abc123def456"
            name:
              type: string
              example: "My App"
            key:
              type: string
              example: "aabpashi_1234567890abcdef..."
              description: The actual API key (only shown on creation)
            userId:
              type: string
              example: "user123"
            permissions:
              type: array
              items:
                type: string
              example: ["read", "write"]
            isActive:
              type: boolean
              example: true
            createdAt:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00.000Z"
            expiresAt:
              type: string
              format: date-time
              example: "2024-02-01T00:00:00.000Z"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        error:
          type: string
          example: "Detailed error information"
...